requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2


#for installation of local packages
-e .